pyproject.toml
[tool.poetry]
name = "djangoskippass"
version = "0.1.0"
description = ""
authors = ["<d.s.puliaiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=3.2.0,<4.0.0"
bson = "^0.5.10"
python-dotenv = "^1.0.1"
pymongo = "^4.6.3"
psycopg2 = "^2.9.9"
djongo = ">=1.3.6,<2.0.0"
sqlparse = ">=0.2.4,<0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
